+++ apps/wallets/quickstart-devkit/CHANGELOG.md
@@ -1,5 +1,51 @@
 # @crossmint/wallets-quickstart-devkit
 
+## 0.1.10
+
+### Patch Changes
+
+- @crossmint/client-sdk-react-ui@1.19.21
+
+## 0.1.9
+
+### Patch Changes
+
+- Updated dependencies [611585d]
+  - @crossmint/client-sdk-react-ui@1.19.20
+
+## 0.1.8
+
+### Patch Changes
+
+- Updated dependencies [660e538]
+  - @crossmint/client-sdk-react-ui@1.19.19
+
+## 0.1.7
+
+### Patch Changes
+
+- @crossmint/client-sdk-react-ui@1.19.18
+
+## 0.1.6
+
+### Patch Changes
+
+- @crossmint/client-sdk-react-ui@1.19.17
+
+## 0.1.5
+
+### Patch Changes
+
+- Updated dependencies [73cbe5b]
+  - @crossmint/client-sdk-react-ui@1.19.16
+
+## 0.1.4
+
+### Patch Changes
+
+- Updated dependencies [8f65e3c]
+  - @crossmint/client-sdk-react-ui@1.19.15
+
 ## 0.1.3
 
 ### Patch Changes