+++ packages/client/signers/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @crossmint/client-signers
 
+## 0.0.16
+
+### Patch Changes
+
+- f27237a: NCS: add export-keys event
+
+## 0.0.15
+
+### Patch Changes
+
+- 660e538: Adapt event interface (NCS)
+
+## 0.0.14
+
+### Patch Changes
+
+- 73cbe5b: NCS: Rename events
+
 ## 0.0.13
 
 ### Patch Changes