+++ packages/client/ui/vanilla-ui/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @crossmint/client-sdk-vanilla-ui
 
+## 1.1.47
+
+### Patch Changes
+
+- Updated dependencies [8e4d900]
+  - @crossmint/common-sdk-base@0.5.1
+  - @crossmint/client-sdk-base@1.5.1
+
+## 1.1.46
+
+### Patch Changes
+
+- Updated dependencies [e597884]
+  - @crossmint/client-sdk-base@1.5.0
+  - @crossmint/common-sdk-base@0.5.0
+
 ## 1.1.45
 
 ### Patch Changes