+++ packages/client/wallets/smart-wallet/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @crossmint/client-sdk-smart-wallet
 
+## 0.3.13
+
+### Patch Changes
+
+- Updated dependencies [8e4d900]
+  - @crossmint/common-sdk-base@0.5.1
+  - @crossmint/client-sdk-base@1.5.1
+
+## 0.3.12
+
+### Patch Changes
+
+- Updated dependencies [e597884]
+  - @crossmint/client-sdk-base@1.5.0
+  - @crossmint/common-sdk-base@0.5.0
+
 ## 0.3.11
 
 ### Patch Changes